create paginatedResponse generic in global types and refactor backend to use
