--- apps/mobile/app/types/apiTypes.ts
@@ -1,7 +1,0 @@
-export interface PaginatedResponse<T> {
-  data: T[];
-  total: number;
-  page: number;
-  limit: number;
-  totalPages: number;
-}
