+++ services/api/src/types/expense.ts
@@ -1,6 +1,6 @@
 import { Expense } from '../../generated/prisma/client';
 export type PaginatedExpenses = {
-  expenses: Expense[];
+  data: Expense[];
   total: number;
   page: number;
   limit: number;
